# src/openapi.yaml
openapi: 3.0.3

info:
  title: Star Wars Explorer API
  version: 0.2.0
  description: |
    API backend para explorar dados de Star Wars a partir da SWAPI.
    Expõe um contrato próprio com paginação, filtros, envelope padrão
    e endpoints correlacionados.

servers:
  - url: https://example-gateway-url
    description: API Gateway
  - url: http://localhost:8080
    description: Local (Functions Framework)

paths:
  /health:
    get:
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeHealth"
              examples:
                ok:
                  value:
                    data:
                      status: ok
                    meta:
                      request_id: "c2f0f6a0-2f2d-4dc5-9f1f-0a7b0b7f2a33"
                    links:
                      self: "/health"
                    errors: []

  /films:
    get:
      summary: List films
      operationId: listFilms
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Films list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeFilmsList"
        "400":
          description: Validation error
        "502":
          description: Upstream error
        "504":
          description: Upstream timeout

  /people:
    get:
      summary: List people
      operationId: listPeople
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: People list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopePeopleList"
        "400":
          description: Validation error
        "502":
          description: Upstream error
        "504":
          description: Upstream timeout

  /planets:
    get:
      summary: List planets
      operationId: listPlanets
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Planets list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopePlanetsList"
        "400":
          description: Validation error
        "502":
          description: Upstream error
        "504":
          description: Upstream timeout

components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Página (1-based).

    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 10
      description: |
        Tamanho da página (1..50).
        Observação: a SWAPI pagina em blocos fixos; este valor
        limita os resultados localmente.

    Search:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Busca textual (mapeada para `search` da SWAPI).

  schemas:
    ErrorItem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true

    Meta:
      type: object
      additionalProperties: false
      required: [request_id]
      properties:
        request_id:
          type: string
        page:
          type: integer
          minimum: 1
          nullable: true
        page_size:
          type: integer
          minimum: 1
          nullable: true
        count:
          type: integer
          minimum: 0
          nullable: true
        total:
          type: integer
          minimum: 0
          nullable: true

    Links:
      type: object
      additionalProperties: false
      required: [self]
      properties:
        self:
          type: string
        next:
          type: string
          nullable: true
        prev:
          type: string
          nullable: true

    HealthData:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]

    Film:
      type: object
      additionalProperties: true
      required: [id, title, url]
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string

    Person:
      type: object
      additionalProperties: true
      required: [id, name, url]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string

    Planet:
      type: object
      additionalProperties: true
      required: [id, name, url]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string

    EnvelopeBase:
      type: object
      additionalProperties: false
      required: [data, meta, links, errors]
      properties:
        data:
          nullable: true
        meta:
          $ref: "#/components/schemas/Meta"
        links:
          $ref: "#/components/schemas/Links"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorItem"

    EnvelopeHealth:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/HealthData"

    EnvelopeFilmsList:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Film"

    EnvelopePeopleList:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Person"

    EnvelopePlanetsList:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Planet"